requirements.txt
pydantic>=2.6
PyYAML>=6.0
jsonschema>=4.23
typing_extensions>=4.9
numpy>=1.26
torch>=2.4
